change memory and registers from usize to u16

+++ src/constants.rs
@@ -15,40 +15,40 @@ pub const R_COND: usize = 9;
 pub const R_COUNT: usize = 10;
 
 
-pub const MR_KBSR: usize = 0xFE00; /* keyboard status */
-pub const MR_KBDR: usize = 0xFE02;
+pub const MR_KBSR: u16 = 0xFE00; /* keyboard status */
+pub const MR_KBDR: u16 = 0xFE02;
 
 
-pub const OP_BR: usize = 0; /* branch */
-pub const OP_ADD: usize = 1;    /* add  */
-pub const OP_LD: usize =  2;     /* load */
-pub const OP_ST: usize =  3;     /* store */
-pub const OP_JSR: usize =  4;    /* jump register */
-pub const OP_AND: usize = 5;  /* bitwise and */
-pub const OP_LDR: usize = 6;    /* load register */
-pub const OP_STR: usize = 7;    /* store register */
-pub const OP_RTI: usize = 8;    /* unused */
-pub const OP_NOT: usize = 9;    /* bitwise not */
-pub const OP_LDI: usize = 10;    /* load indirect */
-pub const OP_STI: usize = 11;    /* store indirect */
-pub const OP_JMP: usize = 12;    /* jump */
-pub const OP_RES: usize = 13;    /* reserved (unused) */
-pub const OP_LEA: usize = 14;    /* load effective address */
-pub const OP_TRAP: usize = 15;    /* execute trap */
+pub const OP_BR: u16 = 0; /* branch */
+pub const OP_ADD: u16 = 1;    /* add  */
+pub const OP_LD: u16 =  2;     /* load */
+pub const OP_ST: u16 =  3;     /* store */
+pub const OP_JSR: u16 =  4;    /* jump register */
+pub const OP_AND: u16 = 5;  /* bitwise and */
+pub const OP_LDR: u16 = 6;    /* load register */
+pub const OP_STR: u16 = 7;    /* store register */
+pub const OP_RTI: u16 = 8;    /* unused */
+pub const OP_NOT: u16 = 9;    /* bitwise not */
+pub const OP_LDI: u16 = 10;    /* load indirect */
+pub const OP_STI: u16 = 11;    /* store indirect */
+pub const OP_JMP: u16 = 12;    /* jump */
+pub const OP_RES: u16 = 13;    /* reserved (unused) */
+pub const OP_LEA: u16 = 14;    /* load effective address */
+pub const OP_TRAP: u16 = 15;    /* execute trap */
 
 
 // 0111000110111111
 
 
-pub const FL_POS: usize = 1 << 0; /* P */
-pub const FL_ZRO: usize = 1 << 1; /* Z */
-pub const FL_NEG: usize = 1 << 2; /* N */
+pub const FL_POS: u16 = 1 << 0; /* P */
+pub const FL_ZRO: u16 = 1 << 1; /* Z */
+pub const FL_NEG: u16 = 1 << 2; /* N */
 
-pub const TRAP_GETC: usize = 0x20;  /* get character from keyboard, not echoed onto the terminal */
-pub const TRAP_OUT: usize = 0x21;   /* output a character */
-pub const TRAP_PUTS: usize = 0x22;  /* output a word string */
-pub const TRAP_IN: usize = 0x23;    /* get character from keyboard, echoed onto the terminal */
-pub const TRAP_PUTSP: usize = 0x24; /* output a byte string */
-pub const TRAP_HALT: usize = 0x25;   /* halt the program */
+pub const TRAP_GETC: u16 = 0x20;  /* get character from keyboard, not echoed onto the terminal */
+pub const TRAP_OUT: u16 = 0x21;   /* output a character */
+pub const TRAP_PUTS: u16 = 0x22;  /* output a word string */
+pub const TRAP_IN: u16 = 0x23;    /* get character from keyboard, echoed onto the terminal */
+pub const TRAP_PUTSP: u16 = 0x24; /* output a byte string */
+pub const TRAP_HALT: u16 = 0x25;   /* halt the program */
